+++ src/CLIEngine/core.cpp
@@ -1,4 +1,4 @@
-#include "CLIEngine_core.hpp"
+#include "CLIEngine/core.hpp"
 
 namespace CLIEngine {
 
@@ -154,6 +154,34 @@ void Sprite::draw(const Coordinate& offset)
     }
     setPalette();
     moveCursor(offset);
+}
+
+Screen::Screen(const std::string& name) : name(name)
+{
+}
+
+void Screen::clear()
+{
+    system("cls");
+}
+
+void Screen::wait()
+{
+    Sleep(1000 / FPS);
+}
+
+ScreenTransition Screen::loop(const ScreenTransition& previous)
+{
+    this->enter(previous);
+
+    std::optional<ScreenTransition> future = std::nullopt;
+    while (!future.has_value()) {
+        future = this->update();
+    }
+
+    this->exit(future.value());
+
+    return future.value();
 }
 
 } // end of namespace CLIEngine
